requirements.txt
mistralai==1.9.11
requests==2.31.0
python-dotenv==1.0.0
